refactor tx details display code, add and use new transaction struct fields
