+++ Decred Wallet/Features/History/TransactionTableViewCell.swift
@@ -10,7 +10,6 @@ import Foundation
 import UIKit
 
 class TransactionTableViewCell: BaseTableViewCell {
-    
     @IBOutlet weak var dataImage: UIImageView!
     @IBOutlet weak var dataText: UILabel!
     @IBOutlet weak var status: UILabel!
@@ -25,27 +24,15 @@ class TransactionTableViewCell: BaseTableViewCell {
     }
     
     override func setData(_ data: Any?) {
-        
         if let transaction = data as? Transaction {
-            let bestBlock =  AppDelegate.walletLoader.wallet?.getBestBlock()
-            var confirmations = 0
-            if(transaction.Height != -1){
-                confirmations = Int(bestBlock!) - transaction.Height
-                confirmations += 1
+            var confirmations: Int32 = 0
+            if transaction.BlockHeight != -1 {
+                confirmations = AppDelegate.walletLoader.wallet!.getBestBlock() - Int32(transaction.BlockHeight) + 1
             }
-
-            if (transaction.Height == -1) {
-                self.status.textColor = UIColor(hex:"#3d659c")
-                self.status.text = LocalizedStrings.pending
-            } else {
-                if (Settings.spendUnconfirmed || confirmations > 1) {
-                    self.status.textColor = UIColor(hex:"#2DD8A3")
-                    self.status.text = LocalizedStrings.confirmed
-                } else {
-                    self.status.textColor = UIColor(hex:"#3d659c")
-                    self.status.text = LocalizedStrings.pending
-                }
-            }
+            
+            let isConfirmed = Settings.spendUnconfirmed || confirmations > 1
+            self.status.text = isConfirmed ? LocalizedStrings.confirmed : LocalizedStrings.pending
+            self.status.textColor = isConfirmed ? UIColor(hex:"#2DD8A3") : UIColor(hex:"#3d659c")
             
             let Date2 = NSDate.init(timeIntervalSince1970: TimeInterval(transaction.Timestamp) )
             let dateformater = DateFormatter()
@@ -56,30 +43,29 @@ class TransactionTableViewCell: BaseTableViewCell {
             dateformater.string(from: Date2 as Date)
             
             self.dateT.text = dateformater.string(from: Date2 as Date)
-            let amount = Decimal(transaction.Amount / 100000000.00) as NSDecimalNumber
             let requireConfirmation = Settings.spendUnconfirmed ? 0 : 2
             
             if (transaction.Type.lowercased() == "regular") {
                 if (transaction.Direction == 0) {
                     let attributedString = NSMutableAttributedString(string: "-")
-                    attributedString.append(Utils.getAttributedString(str: amount.round(8).description, siz: 13.0, TexthexColor: GlobalConstants.Colors.TextAmount))
+                    attributedString.append(Utils.getAttributedString(str: transaction.dcrAmount.round(8).description, siz: 13.0, TexthexColor: GlobalConstants.Colors.TextAmount))
                     self.dataText.attributedText = attributedString
                     self.dataImage?.image = UIImage(named: "debit")
                 } else if(transaction.Direction == 1) {
                     let attributedString = NSMutableAttributedString(string: " ")
-                    attributedString.append(Utils.getAttributedString(str: amount.round(8).description, siz: 13.0, TexthexColor: GlobalConstants.Colors.TextAmount))
+                    attributedString.append(Utils.getAttributedString(str: transaction.dcrAmount.round(8).description, siz: 13.0, TexthexColor: GlobalConstants.Colors.TextAmount))
                     self.dataText.attributedText = attributedString
                     self.dataImage?.image = UIImage(named: "credit")
                 } else if(transaction.Direction == 2) {
                     let attributedString = NSMutableAttributedString(string: " ")
-                    attributedString.append(Utils.getAttributedString(str: amount.round(8).description, siz: 13.0, TexthexColor: GlobalConstants.Colors.TextAmount))
+                    attributedString.append(Utils.getAttributedString(str: transaction.dcrAmount.round(8).description, siz: 13.0, TexthexColor: GlobalConstants.Colors.TextAmount))
                     self.dataText.attributedText = attributedString
                     self.dataImage?.image = UIImage(named: "account")
                 }
             } else if(transaction.Type.lowercased() == "vote") {
                 self.dataText.text = " \(LocalizedStrings.vote)"
                 self.dataImage?.image = UIImage(named: "vote")
-            } else if (transaction.Type.lowercased() == "ticket_purchase") {
+            } else if (transaction.Type.lowercased() == "ticket") {
                 self.dataText.text = " \(LocalizedStrings.ticket)"
                 self.dataImage?.image = UIImage(named: "immature")
                 if (confirmations < requireConfirmation){
