+++ Decred Wallet/Features/TransactionDetails/Cells/TransactionDetailCell.swift
