+++ Decred Wallet/Features/TransactionDetails/Cells/TransactionDetailCell.xib
