+++ Decred Wallet/Features/TransactionDetails/Cells/TransactiontInputDetailsCell.xib
@@ -17,7 +17,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="none" indentationWidth="10" reuseIdentifier="TransactiontInputDetails" rowHeight="268" id="KGk-i7-Jjw" customClass="TransactiontInputDetails" customModule="Decred_Wallet" customModuleProvider="target">
+        <tableViewCell contentMode="scaleToFill" selectionStyle="none" indentationWidth="10" reuseIdentifier="TransactiontInputDetailsCell" rowHeight="268" id="KGk-i7-Jjw" customClass="TransactiontInputDetailsCell" customModule="Decred_Wallet" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="320" height="268"/>
             <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
             <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
