+++ Decred Wallet/Features/TransactionDetails/Cells/TransactiontOutputDetailsCell.swift
@@ -9,72 +9,60 @@
 import UIKit
 
 class TransactiontOutputDetailsCell: UITableViewCell {
-    
     @IBOutlet private weak var viewContainer: UIView!
     @IBOutlet weak var creditsStack: UIStackView!
     @IBOutlet weak var alcCreditStackHeight: NSLayoutConstraint!
-    
+
     var expandOrCollapse: (() -> Void)?
-    
-    var presentingController: TransactionFullDetailsViewController!
-    
-    func setup(with credits:[Credit], decodedOutputs: [DecodedOutput], presentingController: TransactionFullDetailsViewController){
-        creditsStack.subviews.forEach({ $0.removeFromSuperview() }) // this stack view comes with previous items when this function is called again
+
+    var presentingController: TransactionDetailsViewController!
+
+    func setup(_ outputs:[TxOutput], presentingController: TransactionDetailsViewController) {
         self.presentingController = presentingController
         
-        var walletOutputIndices = [Int]()
-        
-        let wallet = AppDelegate.walletLoader.wallet
-        
-        for (_, credit) in credits.enumerated() {
-            walletOutputIndices.append(Int(credit.Index))
+        // this stack view comes with previous items when this function is called again
+        creditsStack.subviews.forEach({ $0.removeFromSuperview() })
+
+        for (_, output) in outputs.enumerated() {
+            var amount = Utils.getAttributedString(
+                str: "\(output.dcrAmount.round(8))",
+                siz: 13,
+                TexthexColor: GlobalConstants.Colors.TextAmount
+            )
+            var address = output.Address
             
-            let amount = Utils.getAttributedString(str: "\(credit.dcrAmount.round(8))", siz: 13, TexthexColor: GlobalConstants.Colors.TextAmount)
-            let title = " (\((wallet?.account(ofAddress: credit.Address))!))"
-            let address = credit.Address
+            var title = output.AccountNumber >= 0 ? output.AccountName: LocalizedStrings.external.lowercased()
+            title = " (\(title))"
             
-            addSubrow(with: amount, title: title, subTitle: address)
-        }
-        
-        for(i, decodedOutput) in decodedOutputs.enumerated() {
-            if walletOutputIndices.contains(i) {
-                continue
-            }
-            
-            var address = decodedOutput.Addresses.count > 0 ? decodedOutput.Addresses[0] : ""
-            var amount = Utils.getAttributedString(str: "\(decodedOutput.dcrAmount.round(8))", siz: 13, TexthexColor: GlobalConstants.Colors.TextAmount)
-            var title = " (\(LocalizedStrings.external.lowercased()))"
-            
-            switch(decodedOutput.ScriptType){
+            switch (output.ScriptType) {
             case "nulldata":
                 amount = NSAttributedString(string: "[\(LocalizedStrings.nullData)]")
                 address = "[\(LocalizedStrings.script)]"
                 title = ""
-                break
+                
             case "stakegen":
                 address = "[\(LocalizedStrings.stakegen)]"
-                break
+                
             default:
                 break
             }
-            
+
             addSubrow(with: amount, title: title, subTitle: address)
         }
-        
+
         // each debit row has an height of 45
         alcCreditStackHeight.constant = CGFloat(45 * creditsStack.arrangedSubviews.count)
     }
-    
+
     @IBAction func hideOrExpandAction(_ sender: UIButton) {
         self.viewContainer.isHidden = false
     }
-    
-    private func addSubrow(with amount: NSAttributedString, title: String, subTitle: String){
-    
+
+    private func addSubrow(with amount: NSAttributedString, title: String, subTitle: String) {
         let subrow = UIView(frame: CGRect(x:0.0, y:0.0, width: self.frame.size.width, height:45.0))
         let amountLabel = UILabel(frame: CGRect(x:5.0, y:1.0, width: self.frame.size.width, height: 22.0))
         let subTitleLabel = UIButton(frame: CGRect(x: 5.0, y: 23, width: self.frame.size.width, height: 22.0))
-        
+
         subTitleLabel.setTitleColor(#colorLiteral(red: 0.2470588235, green: 0.4941176471, blue: 0.8901960784, alpha: 1), for: .normal)
         subTitleLabel.addTarget(self, action: #selector(buttonClicked), for: .touchUpInside)
         subTitleLabel.set(fontSize: 15, name: "SourceSansPro-Regular")
@@ -89,11 +77,11 @@ class TransactiontOutputDetailsCell: UITableViewCell {
         combine.append(NSMutableAttributedString(string: title))
         amountLabel.attributedText = combine
         subTitleLabel.setTitle(subTitle, for: .normal)
-        
+
         creditsStack.addArrangedSubview(subrow)
     }
-    
-    @objc func buttonClicked(sender : UIButton){
+
+    @objc func buttonClicked(sender : UIButton) {
         DispatchQueue.main.async {
             //Copy a string to the pasteboard.
             UIPasteboard.general.string = sender.titleLabel?.text
