+++ Decred Wallet/Features/TransactionDetails/Cells/TransactiontOutputDetailsCell.xib
